requirements.txt
discord.py
google-genai
aiohttp
Pillow
bing-image-downloader
python-dotenv
requests
